A bunch of stuff was started and got muddled up, so some context
The goal is to get a complete, basic end-to-end system running.
This led me to start hooking up things like a path cache (which probably
isn't actually required right away). Then I realized that the
BuildDescription didn't really need to know about rules, but we did need
a basic evaluation environment to keep track of rules.