Rebuilder does mtime checks in build and scheduler loop works.
The rebuilder change was required so we wouldn't have to share the mtime
cache across threads and we can keep mtime checks in the main thread.
This lets us not have any state right now and just pass the rebuilder
the pretend mtime cache.

I've tried rayon and it doesn't have the thread behavior we want.
specifically, we want to send a thread a task, have it run the task and
respond with a result, where the task may be scoped (i.e. have a non
'static lifetime), without needing to have the rebuilder and everything
in a rayon scope. so we may want a simple thread scheduler after all.

Getting the "have these traits be as generic as possible" is also hard.
we may want to just consider having an async build task trait and
allowing schedulers to implement either/or/and for the different kind of
tasks.

I also don't know how to narrow a trait bound in a trait implementation.
i.e. how can a scheduler impl say it accepts rebuilders, but only if
they implement certain traits that the original Scheduler trait did not
define.