Significantly rework lexing to allow parsing literals with expressions.
Previously, the identifier/value/path reading routines would expect
callers to have made progress on the first character. This got awkward
very quickly as every caller had to remember to do it, and sometimes,
when reading expressions, it would cause problems around determining
when to continue looping looking for more literals/escapes.

Changes to a model where all of those routines do not expect to be
advanced past the first character. This now supports escapes and
references in values and yields them as an expression.