Fix a rebuilder bug
Since mark_dirty wasn't considering if a key was multi/single, in a
chain like

foo -> foo.o -> foo.c

if foo.o was touched, foo would still fail to rebuild.

This is because the check for foo.o would notice that foo.o was newer
than foo.c and mark it as Clean (when it should be Modified). That is,
Clean should never overwrite Modified for a single key. Clean only
exists to have a valid cache entry for multi keys.

Single keys -> absent in cache, do not exist, modified, dirty
Multi keys -> absent in cache, dirty, clean