Support for peeking to allow parse_rule to read multiple bindings
Due to limitations on lexer use (we need a Lexer ref to it and not just
an Iterator ref), we can't use Peekable.
Due to how the lexer positioning etc. works, putting peek in the lexer
itself is probably harder (actually, in hindsight, not necessarily). Any
way, adds the ability to peek and then tweaks parse_rule to understand
multiple bindings. Still suffers from the inability to store those
bindings in unevaluated form.