Add support for rule binding evaluation
With this change, build edges evaluate with awareness of bindings
declared in their rules (while the bindings declared in the build have
precedence - only in and out for now).

This involved changing things so that `command` is no longer special in
the parse stage, just another binding. It is specialized and evaluated
in the to_description pass.