Getting there with the trait bounds
Phew! Ok, that required a lot of work, but we have managed to make the
interface for the rebuilder and scheduler a little more in line with the
paper.

The rebuilder now has the error as an associated type as it seems
more idiomatic rust to have "output types" be associated types.
Rebuilders are required to return tasks, instead of optional.
This is also nice in the ninja-specific case because it fixes our
outstanding jobs count to be in-line with ninja OG.

The scheduler is liberated from being generic on the rebuilder error.
This does cause problems in that I haven't figured out how to restrict
the rebuilder such that a rebuilder is only accepted if its error can be
converted to a builder error.

Removed the State parameter for now because we have enough type system
problems on our hands without it and we don't need it right now :P