Take starting keys into account when building the graph.
Apart from doing less work, this has the benefit of fixing a bug.
Previously, we would build a graph of all tasks, regardless of if all
tasks were going to be built. This meant it was possible for keys to end
up not in finished, not in waiting and not in ready because they were
never reachable from the set of keys requested. This no longer happens.